Move clean_generated_code to utils
