+++ pyproject.toml
@@ -8,7 +8,7 @@ dependencies = [
     "accelerate>=1.10.0",
     "alpaca-eval>=0.6.6",
     "anthropic>=0.62.0",
-    "bitsandbytes>=0.47.0",
+    "bitsandbytes>=0.47.0 ; platform_system != 'Darwin'",
     "datasets==3.6.0",
     "inspect-ai==0.3.120",
     "matplotlib>=3.10.5",
