--- scripts/training/trl-test.py
@@ -1,206 +1,0 @@
-# %%
-import torch
-from peft import LoraConfig, get_peft_model
-from transformers import AutoModelForCausalLM, AutoProcessor, BitsAndBytesConfig
-import re
-
-from datasets import Dataset, load_dataset
-
-ckpt = "meta-llama/Llama-3.1-8B-Instruct"
-
-# Set up quantization with recommended and explicit settings for stability and performance
-quantization_config = BitsAndBytesConfig(
-    load_in_4bit=True,
-    bnb_4bit_use_double_quant=True,  # Improves quantization accuracy
-    bnb_4bit_quant_type="nf4",  # Use NormalFloat4 for better performance
-    bnb_4bit_compute_dtype=torch.bfloat16,  # Match model dtype for efficiency
-)
-
-model = AutoModelForCausalLM.from_pretrained(
-    ckpt,
-    device_map="auto",
-    torch_dtype=torch.bfloat16,
-    attn_implementation="eager",
-    quantization_config=quantization_config,
-)
-
-# Load LoRA with reduced rank to save memory
-lora_config = LoraConfig(
-    task_type="CAUSAL_LM",
-    r=8,  # Reduced from 8 to 4
-    lora_alpha=32,  # Reduced from 32 to 16
-    target_modules="all-linear",
-    # Add dropout to reduce memory usage during training
-    lora_dropout=0.1,
-)
-model = get_peft_model(model, lora_config)
-print(model.print_trainable_parameters())
-
-processor = AutoProcessor.from_pretrained(ckpt)
-
-# %%
-
-# Load and prep dataset
-SYSTEM_PROMPT = """
-Respond in the following format:
-<reasoning>
-...
-</reasoning>
-<answer>
-...
-</answer>
-"""
-
-XML_COT_FORMAT = """\
-<reasoning>
-{reasoning}
-</reasoning>
-<answer>
-{answer}
-</answer>
-"""
-
-
-def extract_xml_answer(text: str) -> str:
-    answer = text.split("<answer>")[-1]
-    answer = answer.split("</answer>")[0]
-    return answer.strip()
-
-
-def extract_hash_answer(text: str) -> str | None:
-    if "####" not in text:
-        return None
-    return text.split("####")[1].strip()
-
-
-# uncomment middle messages for 1-shot prompting
-def get_gsm8k_questions(split="train") -> Dataset:
-    data = load_dataset("openai/gsm8k", "main")[split]  # type: ignore
-    data = data.map(
-        lambda x: {  # type: ignore
-            "prompt": [
-                {"role": "system", "content": SYSTEM_PROMPT},
-                {"role": "user", "content": x["question"]},
-            ],
-            "answer": extract_hash_answer(x["answer"]),
-        }
-    )  # type: ignore
-    return data  # type: ignore
-
-
-dataset = get_gsm8k_questions()
-
-
-# %%
-def correctness_reward_func(prompts, completions, answer, **kwargs) -> list[float]:
-    responses = [completion[0]["content"] for completion in completions]
-    q = prompts[0][-1]["content"]
-    extracted_responses = [extract_xml_answer(r) for r in responses]
-    print(
-        "-" * 20,
-        f"Question:\n{q}",
-        f"\nAnswer:\n{answer[0]}",
-        f"\nResponse:\n{responses[0]}",
-        f"\nExtracted:\n{extracted_responses[0]}",
-    )
-    return [2.0 if r == a else 0.0 for r, a in zip(extracted_responses, answer)]
-
-
-def int_reward_func(completions, **kwargs) -> list[float]:
-    responses = [completion[0]["content"] for completion in completions]
-    extracted_responses = [extract_xml_answer(r) for r in responses]
-    return [0.5 if r.isdigit() else 0.0 for r in extracted_responses]
-
-
-def strict_format_reward_func(completions, **kwargs) -> list[float]:
-    """Reward function that checks if the completion has a specific format."""
-    pattern = r"^<reasoning>\n.*?\n</reasoning>\n<answer>\n.*?\n</answer>\n$"
-    responses = [completion[0]["content"] for completion in completions]
-    matches = [re.match(pattern, r) for r in responses]
-    return [0.5 if match else 0.0 for match in matches]
-
-
-def soft_format_reward_func(completions, **kwargs) -> list[float]:
-    """Reward function that checks if the completion has a specific format."""
-    pattern = r"<reasoning>.*?</reasoning>\s*<answer>.*?</answer>"
-    responses = [completion[0]["content"] for completion in completions]
-    matches = [re.match(pattern, r) for r in responses]
-    return [0.5 if match else 0.0 for match in matches]
-
-
-def count_xml(text) -> float:
-    count = 0.0
-    if text.count("<reasoning>\n") == 1:
-        count += 0.125
-    if text.count("\n</reasoning>\n") == 1:
-        count += 0.125
-    if text.count("\n<answer>\n") == 1:
-        count += 0.125
-        count -= len(text.split("\n</answer>\n")[-1]) * 0.001
-    if text.count("\n</answer>") == 1:
-        count += 0.125
-        count -= (len(text.split("\n</answer>")[-1]) - 1) * 0.001
-    return count
-
-
-def xmlcount_reward_func(completions, **kwargs) -> list[float]:
-    contents = [completion[0]["content"] for completion in completions]
-    return [count_xml(c) for c in contents]
-
-
-# %%
-from trl import GRPOConfig, GRPOTrainer
-
-# Reduced sequence lengths to save memory
-max_prompt_length = 128  # Reduced from 256
-max_seq_length = 512  # Reduced from 1024
-
-
-training_args = GRPOConfig(
-    learning_rate=5e-6,
-    adam_beta1=0.9,
-    adam_beta2=0.99,
-    weight_decay=0.1,
-    warmup_ratio=0.1,
-    lr_scheduler_type="cosine",
-    optim="adamw_8bit",
-    logging_steps=1,
-    # Reduced batch size and increased gradient accumulation for memory efficiency
-    per_device_train_batch_size=2,  # Reduced from 2
-    gradient_accumulation_steps=1,  # Increased from 1 (effective batch size = 1 * 4 = 4)
-    # Reduced number of generations to save memory
-    num_generations=2,  # Reduced from 2
-    max_prompt_length=max_prompt_length,
-    max_completion_length=max_seq_length - max_prompt_length,
-    num_train_epochs=1,
-    max_steps=250,
-    save_steps=250,
-    max_grad_norm=0.1,
-    report_to="none",
-    # Add memory optimization settings
-    dataloader_pin_memory=False,  # Disable pin memory to save RAM
-    remove_unused_columns=False,  # Keep unused columns to avoid processing overhead
-)
-
-# %%
-processor.pad_token_id = 0
-processor.bos_token_id = 1
-processor.eos_token_id = 2
-
-trainer = GRPOTrainer(
-    model=model,
-    processing_class=processor,
-    reward_funcs=[
-        xmlcount_reward_func,
-        soft_format_reward_func,
-        strict_format_reward_func,
-        int_reward_func,
-        correctness_reward_func,
-    ],
-    args=training_args,
-    train_dataset=dataset,
-)
-
-trainer.train()
-
-# %%
