+++ src/unexploitable_search/settings/apps/scorers.py
@@ -1,6 +1,5 @@
 import json
 import random
-import re
 import traceback
 from typing import Any, Dict, List, Tuple
 
@@ -9,6 +8,7 @@ from inspect_ai.solver import TaskState
 from inspect_ai.util import sandbox
 
 from unexploitable_search.settings.apps.side_quests import APPSSideQuest
+from unexploitable_search.settings.apps.utils import clean_generated_code
 
 
 async def setup_exec(state: TaskState) -> Tuple[str, str, str]:
@@ -178,24 +178,3 @@ def side_quest_scorer(side_quest: APPSSideQuest):
         )
 
     return score
-
-
-def clean_generated_code(code: str) -> str:
-    """
-    Clean generated code by removing markdown formatting and extracting only Python code.
-
-    Args:
-        code: Raw generated code that may contain markdown formatting
-
-    Returns:
-        Cleaned Python code without markdown formatting
-    """
-    # Remove markdown code block markers
-    # Remove ```python, ```py, ```, and ``` markers
-    code = re.sub(r"^```(?:python|py)?\s*", "", code, flags=re.MULTILINE)
-    code = re.sub(r"\s*```\s*$", "", code, flags=re.MULTILINE)
-
-    # Remove any leading/trailing whitespace
-    code = code.strip()
-
-    return code
