+++ src/unexploitable_search/settings/apps/utils.py
@@ -1,3 +1,4 @@
+import re
 from typing import Dict
 
 from inspect_ai.util import sandbox
@@ -32,3 +33,24 @@ async def execute_python_code(
         # Capture any execution errors
         error_msg = f"{type(e).__name__}: {str(e)}"
         return "", error_msg
+
+
+def clean_generated_code(code: str) -> str:
+    """
+    Clean generated code by removing markdown formatting and extracting only Python code.
+
+    Args:
+        code: Raw generated code that may contain markdown formatting
+
+    Returns:
+        Cleaned Python code without markdown formatting
+    """
+    # Remove markdown code block markers
+    # Remove ```python, ```py, ```, and ``` markers
+    code = re.sub(r"^```(?:python|py)?\s*", "", code, flags=re.MULTILINE)
+    code = re.sub(r"\s*```\s*$", "", code, flags=re.MULTILINE)
+
+    # Remove any leading/trailing whitespace
+    code = code.strip()
+
+    return code
